+++ docs/src/main/kotlin/Main.kt
@@ -1,4 +1,5 @@
 import io.koalaql.docs.examples.quickExample
+import io.koalaql.docs.executing.resultSets
 import io.koalaql.markout.docusaurus.docusaurus
 import io.koalaql.markout.markout
 import kotlin.io.path.Path
@@ -29,6 +30,8 @@ fun main() = markout(Path("site/docs")) {
 
         directory("executing") {
             label = "Executing SQL"
+
+            markdown("results") { resultSets() }
         }
 
         directory("examples") {
